Refactoring SeenFortWorker move logic into MoveToFortWorker (#556)
